Decomposeed Builder Ingredients to smaller components
